Events based cache code reorg
